requirements.txt
scipy
cython
scikit-learn
iso8601
torchvision
tqdm
pandas
pydantic
pyyaml
tabulate
numba
typing_extensions
wandb
plotly
h5py
pytorch-lightning
scikit-image
lightning-bolts
matplotlib
seaborn
